Extract adaptivePoll into shared perfuncted/poll package
Create poll.AdaptivePoll with exponential backoff (base * 2^attempt,
capped at max). Update find package to import pollpkg instead of keeping
a local copy. Move test from find package into poll package.
